Dash Bugfixes (#170)
* tmp

tmp

works

set adc fault

set fault

scale throttle

bruh

Bugs

Throttle bugfixes from FCTT

gps -> amk speed

dash changes

- overload
- pumps on race
- igbt

shock pots working

bad code timebar

* RPM to Speed

* gps speed

* temp

* shocks

* constant refresh patch

* tv25 page update

* TV integration

* Shock calibration

* send can message

* Final Calibration changes

* pdu: Adjust cooling


* cooling small edits

* fixed GPS stale bug

* small tweak

* fan

* renable lvsense

* calibratiob pt2

---------

+++ source/dashboard/lcd/menu_system.h
@@ -1,10 +1,10 @@
 /**
  * @file menu_system.h
  * @brief Implementation of menu system for LCD display interface
- * 
+ *
  * Provides functions for managing menu navigation, element styling,
  * and user interaction with the LCD display menu system.
- * 
+ *
  * @author Irving Wang (wang5952@purdue.edu)
  */
 
@@ -37,9 +37,9 @@ typedef struct {
     element_state_t state;
     char* object_name;          // Nextion object name
     void (*on_change)(void);    // Callback when value changes
-    uint8_t current_value;      // Current value for numeric types or on/off state for toggles
-    uint8_t min_value;          // Minimum value for numeric types
-    uint8_t max_value;          // Maximum value for numeric types
+    uint16_t current_value;     // Current value for numeric types or on/off state for toggles
+    uint16_t min_value;         // Minimum value for numeric types
+    uint16_t max_value;         // Maximum value for numeric types
     uint8_t increment;          // Increment for numeric types
 } menu_element_t;
